requirements.txt
flask>=2.3.0
requests>=2.28.0
pyyaml>=6.0
colorama>=0.4.6
prometheus-client>=0.17.0
schedule>=1.2.0
psutil>=5.9.0
pytz>=2023.3
click==8.1.7
tabulate==0.9.0
validators==0.22.0 